refactor: extract product details hook
